feat: evm deposit and call command
